Moves views to the UI app.
closes #108
