--- ui/templates/status.html
@@ -1,14 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-<p>
-Number of repositories: {{ repos }}
-</p>
-
-<p>
-Number of courses: {{ courses }}
-</p>
-
-<p>
-Number of learning resources: {{ resources }}
-</p>
-{% endblock %}
