+++ ui/urls.py
@@ -19,9 +19,10 @@ from __future__ import unicode_literals
 from django.conf.urls import include, url
 from django.contrib import admin
 
-from learningresources.views import welcome, create_repo, listing, export
-from importer.views import status, upload
-from taxonomy.views import edit_vocabulary, create_vocabulary
+from ui.views import (
+    welcome, create_repo, listing, export,
+    upload, edit_vocabulary, create_vocabulary,
+)
 import cas.urls as cas_urls
 
 urlpatterns = [
@@ -32,7 +33,6 @@ urlpatterns = [
     url(r'^repositories/(?P<repo_slug>[-\w]+)/$', listing,
         name='listing'),
     url(r'^learningresources/(?P<resource_id>\d+)/$', export, name='export'),
-    url(r'^status/$', status, name='status'),
     url(r'^repositories/(?P<repo_slug>[-\w]+)/import/$',
         upload, name='upload'),
     url(r'^cas/', include(cas_urls)),
